Move mocks to folder with files so watch works
