+++ src/applications/vaos/services/mocks/index.js
@@ -1,4 +1,6 @@
 /* istanbul ignore file */
+/* eslint-disable camelcase */
+const delay = require('mocker-api/lib/delay');
 
 // var
 const confirmedVA = require('./var/confirmed_va.json');
@@ -25,19 +27,7 @@ const generateMockSlots = require('./var/slots.js');
 
 varSlots.data[0].attributes.appointmentTimeSlot = generateMockSlots();
 
-/*
- * Handler definition:
- *
- * path: Can be a string or a regex. Strings match from the end of a url
- * delay: Value in ms to delay the response by
- * response: Can be a json object or a function
- * 
- * response function params:
- * url: the full url of the request
- * options.requestData: Parsed body data
- * options.groups: Any matched regex groups from the url
- */
-module.exports = {
+const responses = {
   'GET /vaos/v0/appointments': (req, res) => {
     if (req.params.type === 'cc') {
       return res.json(confirmedCC);
@@ -181,4 +171,96 @@ module.exports = {
     },
   },
   'PUT /vaos/v0/preferences': { data: { attributes: {} } },
+  'GET /v0/user': {
+    data: {
+      attributes: {
+        profile: {
+          sign_in: {
+            service_name: 'idme',
+          },
+          email: 'fake@fake.com',
+          loa: { current: 3 },
+          first_name: 'Jane',
+          middle_name: '',
+          last_name: 'Doe',
+          gender: 'F',
+          birth_date: '1985-01-01',
+          verified: true,
+        },
+        veteran_status: {
+          status: 'OK',
+          is_veteran: true,
+          served_in_military: true,
+        },
+        in_progress_forms: [],
+        prefills_available: ['21-526EZ'],
+        services: [
+          'facilities',
+          'hca',
+          'edu-benefits',
+          'evss-claims',
+          'form526',
+          'user-profile',
+          'health-records',
+          'rx',
+          'messaging',
+        ],
+        va_profile: {
+          status: 'OK',
+          birth_date: '19511118',
+          family_name: 'Hunter',
+          gender: 'M',
+          given_names: ['Julio', 'E'],
+          active_status: 'active',
+          facilities: [
+            {
+              facility_id: '983',
+              is_cerner: false,
+            },
+            {
+              facility_id: '984',
+              is_cerner: false,
+            },
+          ],
+        },
+      },
+    },
+    meta: { errors: null },
+  },
+
+  'OPTIONS /v0/maintenance_windows': 'OK',
+  'GET /v0/maintenance_windows': { data: [] },
+
+  'GET /v0/feature_toggles': {
+    data: {
+      type: 'feature_toggles',
+      features: [
+        { name: 'facilityLocatorShowCommunityCares', value: true },
+        { name: 'profile_show_profile_2.0', value: false },
+        { name: 'vaOnlineScheduling', value: true },
+        { name: 'vaOnlineSchedulingCancel', value: true },
+        { name: 'vaOnlineSchedulingRequests', value: true },
+        { name: 'vaOnlineSchedulingCommunityCare', value: true },
+        { name: 'vaOnlineSchedulingDirect', value: true },
+        { name: 'vaOnlineSchedulingPast', value: true },
+        { name: 'vaOnlineSchedulingExpressCare', value: true },
+        { name: 'vaOnlineSchedulingExpressCareNew', value: true },
+        { name: 'vaOnlineSchedulingFlatFacilityPage', value: true },
+        { name: 'vaOnlineSchedulingProviderSelection', value: true },
+        { name: 'vaOnlineSchedulingCheetah', value: true },
+        { name: 'vaOnlineSchedulingHomepageRefresh', value: true },
+        { name: 'vaOnlineSchedulingUnenrolledVaccine', value: true },
+        { name: 'vaGlobalDowntimeNotification', value: false },
+        { name: 'ssoe', value: true },
+        { name: 'ssoeInbound', value: false },
+        { name: 'ssoeEbenefitsLinks', value: false },
+        { name: 'edu_section_103', value: true },
+        { name: 'form526OriginalClaims', value: false },
+        { name: 'vaViewDependentsAccess', value: false },
+        { name: 'gibctEybBottomSheet', value: true },
+      ],
+    },
+  },
 };
+
+module.exports = delay(responses, 1000);
